fix(bedrock): align converse shape and add live smoke tests
